Flavour-specific requirementschecker + associated DI module
