+++ project/app/src/gms/java/org/owntracks/android/gms/location/LatLng.kt
@@ -1,0 +1,7 @@
+package org.owntracks.android.gms.location
+
+import org.owntracks.android.location.LatLng
+
+fun LatLng.toGMSLatLngConversion(): com.google.android.gms.maps.model.LatLng {
+    return com.google.android.gms.maps.model.LatLng(latitude, longitude)
+}
