--- project/app/src/gms/java/org/owntracks/android/location/gms/LatLng.kt
@@ -1,7 +1,0 @@
-package org.owntracks.android.location.gms
-
-import org.owntracks.android.location.LatLng
-
-fun LatLng.toGMSLatLngConversion(): com.google.android.gms.maps.model.LatLng {
-    return com.google.android.gms.maps.model.LatLng(latitude, longitude)
-}
