+++ project/app/src/main/java/org/owntracks/android/injection/modules/SingletonModule.kt
@@ -19,7 +19,6 @@ class SingletonModule {
     }
 
     @Provides
-    @Singleton
     fun provideContactsRepo(eventBus: EventBus?, contactImageProvider: ContactImageProvider?): ContactsRepo {
         return MemoryContactsRepo(eventBus!!, contactImageProvider!!)
     }
