+++ project/app/src/main/java/org/owntracks/android/location/LatLng.kt
@@ -1,6 +1,6 @@
 package org.owntracks.android.location
 
-import org.owntracks.android.location.gms.toGMSLatLngConversion
+import org.owntracks.android.gms.location.toGMSLatLngConversion
 
 data class LatLng(val latitude: Double, val longitude: Double) {
     fun toGMSLatLng(): com.google.android.gms.maps.model.LatLng {
