+++ project/app/src/main/java/org/owntracks/android/location/NoopGeofencingClient.kt
