+++ project/app/src/main/java/org/owntracks/android/location/geofencing/Geofence.kt
